+++ src/Components/HomePage.tsx
@@ -1,1 +1,7 @@
-export const HomePage = () => <h1 className="title">Home page</h1>;
+export const HomePage = () => (
+  <div className="section">
+    <div className="container">
+      <h1 className="title">Home page</h1>
+    </div>
+  </div>
+);
