+++ src/Components/PageNotFound.tsx
@@ -1,1 +1,7 @@
-export const PageNotFound = () => <h1 className="title">Page not found</h1>;
+export const PageNotFound = () => (
+  <div className="section">
+    <div className="container">
+      <h1 className="title">Page not found</h1>
+    </div>
+  </div>
+);
